requirements.txt
numpy==1.19.1
scipy~=1.5.2
xmltodict==0.12.0
icecream
python-dateutil